refactor: Optimize code structure.
