+++ TYLDDB/TYLDDB.cs
@@ -4,9 +4,6 @@ using System.Threading.Tasks;
 using TYLDDB.Basic.Exception;
 using TYLDDB.Parser;
 using TYLDDB.Utils;
-using TYLDDB.Utils.Database;
-using TYLDDB.Utils.FastCache.ConcurrentDictionary;
-using TYLDDB.Utils.FastCache.SemaphoreThreadLock;
 
 namespace TYLDDB
 {
@@ -14,66 +11,8 @@ namespace TYLDDB
     /// The core class of the database.<br />
     /// 数据库的核心类。
     /// </summary>
-    public class LDDB
+    public partial class LDDB
     {
-        /// <summary>
-        /// Instantiate the LDDB class<br />
-        /// 实例化LDDB类
-        /// </summary>
-        public LDDB()
-        {
-            // 实例化数据库操作类
-            database_v1 = new Database_V1();
-            database_v2 = new Database_V2();
-
-            // 实例化并发词典
-            cdStringDictionary = new CdStringDictionary();
-            cdShortDictionary = new CdShortDictionary();
-            cdLongDictionary = new CdLongDictionary();
-            cdIntegerDictionary = new CdIntegerDictionary();
-            cdFloatDictionary = new CdFloatDictionary();
-            cdDoubleDictionary = new CdDoubleDictionary();
-            cdDecimalDictionary = new CdDecimalDictionary();
-            cdCharDictionary = new CdCharDictionary();
-            cdBooleanDictionary = new CdBooleanDictionary();
-
-            // 实例化信号词典
-            stlStringDictionary = new StlStringDictionary();
-            stlShortDictionary = new StlShortDictionary();
-            stlLongDictionary = new StlLongDictionary();
-            stlIntegerDictionary = new StlIntegerDictionary();
-            stlFloatDictionary = new StlFloatDictionary();
-            stlDoubleDictionary = new StlDoubleDictionary();
-            stlDecimalDictionary = new StlDecimalDictionary();
-            stlCharDictionary = new StlCharDictionary();
-            stlBooleanDictionary = new StlBooleanDictionary();
-        }
-
-        ///////////////////////////////////////////////////// 私有字段
-        private string _filePath;  // 存储文件路径
-        private string _fileContent; // 存储文件内容
-        private string _databaseContent; // 存储数据库内容
-        private Database_V1 database_v1;
-        private Database_V2 database_v2;
-        private CdStringDictionary cdStringDictionary;
-        private CdShortDictionary cdShortDictionary;
-        private CdLongDictionary cdLongDictionary;
-        private CdIntegerDictionary cdIntegerDictionary;
-        private CdFloatDictionary cdFloatDictionary;
-        private CdDoubleDictionary cdDoubleDictionary;
-        private CdDecimalDictionary cdDecimalDictionary;
-        private CdCharDictionary cdCharDictionary;
-        private CdBooleanDictionary cdBooleanDictionary;
-        private StlStringDictionary stlStringDictionary;
-        private StlShortDictionary stlShortDictionary;
-        private StlLongDictionary stlLongDictionary;
-        private StlIntegerDictionary stlIntegerDictionary;
-        private StlFloatDictionary stlFloatDictionary;
-        private StlDoubleDictionary stlDoubleDictionary;
-        private StlDecimalDictionary stlDecimalDictionary;
-        private StlCharDictionary stlCharDictionary;
-        private StlBooleanDictionary stlBooleanDictionary;
-
         ///////////////////////////////////////////////////// 公开字段
         /// <summary>
         /// Set the path where you want to read the file<br/>
@@ -99,6 +38,8 @@ namespace TYLDDB
         /// 当前文件内所有数据库的名称
         /// </summary>
         public List<string> AllDatabaseName;
+
+        ///////////////////////////////////////////////////// 方法
 
         /// <summary>
         /// 验证文件路径是否为null或空
